added responsive tablet for our family main page
