+++ src/app/our-family/page.tsx
@@ -1,8 +1,6 @@
 "use client";
 import "./OurFamily.scss";
 import dynamic from "next/dynamic";
-import ChevronDown from "#/assets/svg/chevron-down";
-import SparklesOurFamily from "#/assets/svg/sparkles/sparkles-our-family";
 import HeaderSection from "#/components/ui/headerSection/HeaderSection";
 import SparklesStardustValley from "#/assets/svg/sparkles/sparkles-stardust-valley";
 import StardustValley from "#/components/our-family/stardus-valley/StardustValley";
@@ -19,159 +17,13 @@ import ShiabunIcon from "#/assets/svg/our-family/shiabun-icon";
 import TogetherExcl1 from "#/assets/svg/our-family/together-excl-1";
 import TogetherExcl2 from "#/assets/svg/our-family/together-excl-2";
 import TogetherExcl3 from "#/assets/svg/our-family/together-excl-3";
-import BannerExcl1 from "#/assets/svg/our-family/banner-excl-1";
-import BannerExcl2 from "#/assets/svg/our-family/banner-excl-2";
-import ShiabunSilhouette from "#/assets/svg/our-family/shiabun-silhouette";
-import RoseSilhouette from "#/assets/svg/our-family/rose-silhouette";
-import AicandiiSilhouette from "#/assets/svg/our-family/aicandii-silhouette";
-import BuffpupSilhouette from "#/assets/svg/our-family/buffpup-silhouette";
-import House from "#/assets/svg/house";
-import { motion, useMotionValue, useSpring, useTransform } from "framer-motion";
-import { useState, useEffect } from "react";
-import { bouncyTransition, springTransition } from "#/helpers/const-animations";
+import OurFamilyBanner from "#/components/our-family/banner/OurFamilyBanner";
 const ReactPlayer = dynamic(() => import("react-player"), { ssr: false });
 
 const OurFamily = () => {
-  const [mousePos, setMousePos] = useState({ x: 0, y: 0 });
-  const [windowDimensions, setWindowDimensions] = useState({
-    width: 0,
-    height: 0,
-  });
-
-  const x = useMotionValue(0);
-  const y = useMotionValue(0);
-
-  const smoothX = useSpring(x, { stiffness: 100, damping: 14 });
-  const smoothY = useSpring(y, { stiffness: 100, damping: 14 });
-  const smoothMembersX = useSpring(x, { stiffness: 300, damping: 30 });
-  const smoothMembersY = useSpring(y, { stiffness: 300, damping: 30 });
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      const handleResize = () => {
-        setWindowDimensions({
-          width: window.innerWidth,
-          height: window.innerHeight,
-        });
-      };
-
-      handleResize();
-      window.addEventListener("resize", handleResize);
-
-      return () => {
-        window.removeEventListener("resize", handleResize);
-      };
-    }
-  }, []);
-
-  const houseX = useTransform(smoothX, [0, windowDimensions?.width], [-5, 5]);
-  const houseY = useTransform(smoothY, [0, windowDimensions?.height], [-5, 5]);
-  const membersX = useTransform(
-    smoothMembersX,
-    [0, windowDimensions?.width],
-    [-15, 15]
-  );
-  const membersY = useTransform(
-    smoothMembersY,
-    [0, windowDimensions?.height],
-    [-15, 15]
-  );
-
-  const exclMarkX = useTransform(
-    smoothX,
-    [0, windowDimensions?.width],
-    [-5, 5]
-  );
-  const exclMarkY = useTransform(
-    smoothY,
-    [0, windowDimensions?.height],
-    [-5, 5]
-  );
-
-  const handleMouse = (event: React.MouseEvent) => {
-    const { clientX, clientY } = event;
-    setMousePos({ x: clientX, y: clientY });
-    x.set(clientX);
-    y.set(clientY);
-  };
-
-  const familyDescription = () => {
-    return (
-      <>
-        <span>Wait, but really.</span>
-        <span className="stars">
-          <span>Buffpup</span>, <span>AiCandii</span>, <span>rosedoodle</span>,
-          and <span>ShiaBun</span> — each came from a very different part of
-          Stardust Valley, some even hailing from outside the city proper. But
-          while each had their responsibilities, their hopes, and their goals…
-          Something was missing.
-        </span>
-      </>
-    );
-  };
-
-  const [isHover, setIsHover] = useState(false);
-
   return (
-    <div className="family" onMouseMove={handleMouse}>
-      <div className="family__banner">
-        <HeaderSection
-          as="header"
-          title="OUR FAMILY."
-          preDescription="Four virtual content creators walk into a cafe..."
-          description={familyDescription()}
-          SparklesComponent={SparklesOurFamily}
-          customStyles="our-family"
-        />
-        {/* Background images tracking mouse movement */}
-        <motion.div
-          className="excl-marks"
-          style={{ x: exclMarkX, y: exclMarkY }}
-        >
-          <BannerExcl1 className="excl-1" />
-          <BannerExcl2 className="excl-2" />
-        </motion.div>
-        <motion.div className="members" style={{ x: membersX, y: membersY }}>
-          <ShiabunSilhouette className="shia" />
-          <RoseSilhouette className="rose" />
-          <AicandiiSilhouette className="candii" />
-          <BuffpupSilhouette className="buff" />
-        </motion.div>
-
-        <motion.div className="house" style={{ x: houseX, y: houseY }}>
-          <House />
-        </motion.div>
-
-        <button className="lore-btn">
-          <motion.div
-            className="more-lore"
-            onHoverStart={() => setIsHover(true)}
-            onHoverEnd={() => setIsHover(false)}
-          >
-            <motion.div
-              className="chevron"
-              animate={{
-                transform: isHover ? "translateY(1rem)" : "translateY(0rem)",
-              }}
-              initial={{ transform: "translateY(0rem)" }}
-              transition={springTransition}
-            >
-              <ChevronDown />
-            </motion.div>
-            See more lore
-            <motion.div
-              className="chevron"
-              animate={{
-                transform: isHover ? "translateY(1rem)" : "translateY(0rem)",
-              }}
-              initial={{ transform: "translateY(0rem)" }}
-              transition={springTransition}
-            >
-              <ChevronDown />
-            </motion.div>
-          </motion.div>
-        </button>
-      </div>
+    <div className="family">
+      <OurFamilyBanner />
       <div className="family__stardust-map">
         <div className="header-and-map">
           <HeaderSection
@@ -194,6 +46,8 @@ const OurFamily = () => {
           <ReactPlayer
             url={`https://www.youtube.com/embed/4CxlS6IYnFM&t?autoplay=1&rel=0&modestbranding=1&playsinline=1&color=white`}
             controls={true}
+            width="100%"
+            height="100%"
             className="video"
           />
         </div>
