+++ src/components/ui/headerSection/HeaderSection.scss
@@ -61,7 +61,7 @@
     text-align: center;
     padding: 0 12.8rem;
 
-    @media screen and (max-width: 810px) {
+    @media screen and (max-width: 1320px) {
       font-size: 1.6rem;
       padding: 0;
     }
@@ -277,7 +277,7 @@
   &.our-family {
     padding: 26rem 6.4rem 9.6rem;
 
-    @media screen and (max-width: 810px) {
+    @media screen and (max-width: 1320px) {
       padding: 16rem 2rem 9.6rem;
     }
 
@@ -285,7 +285,7 @@
       line-height: 110%;
     }
     .preDescription {
-      @media screen and (max-width: 810px) {
+      @media screen and (max-width: 1320px) {
         font-size: 2rem;
         padding-bottom: 3rem;
       }
@@ -294,6 +294,11 @@
     .description {
       padding: 0 6.4rem;
       gap: 2rem;
+
+      @media screen and (max-width: 1320px) {
+        padding: 1rem 5rem 0;
+        text-align: center;
+      }
       @media screen and (max-width: 810px) {
         padding: 0;
       }
@@ -317,7 +322,12 @@
       top: -2rem;
       left: 50%;
       transform: translateX(-51%);
-
+      @media screen and (max-width: 1320px) {
+        scale: 0.7;
+        top: -3rem;
+        left: 0%;
+        transform: translateX(-21%);
+      }
       @media screen and (max-width: 810px) {
         scale: 0.5;
         top: -4rem;
@@ -329,6 +339,7 @@
 
   &.stardust-valley {
     padding: 12.8rem 6.4rem 0rem;
+
     @media screen and (max-width: 810px) {
       padding: 12.8rem 0rem 0rem;
     }
@@ -337,10 +348,13 @@
       line-height: 110%;
       font-size: 9.6rem;
 
-      @media screen and (max-width: 810px) {
-        font-size: 4rem;
+      @media screen and (max-width: 1320px) {
+        font-size: 6.4rem;
         white-space: nowrap;
         margin: 0 auto;
+      }
+      @media screen and (max-width: 810px) {
+        font-size: 4rem;
       }
     }
     .sparkles {
@@ -348,6 +362,13 @@
       top: -1rem;
       left: 50%;
       transform: translateX(-50%);
+
+      @media screen and (max-width: 1320px) {
+        scale: 0.7;
+        top: -2rem;
+        left: 0%;
+        transform: translateX(-25%);
+      }
 
       @media screen and (max-width: 810px) {
         scale: 0.5;
@@ -366,9 +387,12 @@
       font-size: 9.6rem;
       margin-bottom: 1.5rem;
 
+      @media screen and (max-width: 1320px) {
+        font-size: 6.4rem;
+        white-space: nowrap;
+      }
       @media screen and (max-width: 810px) {
         font-size: 4rem;
-        white-space: nowrap;
       }
     }
     .description {
@@ -382,6 +406,14 @@
       top: -3rem;
       left: 50%;
       transform: translateX(-51%);
+
+      @media screen and (max-width: 1320px) {
+        scale: 0.7;
+        top: -5rem;
+        left: 0%;
+        transform: translateX(-27%);
+      }
+
       @media screen and (max-width: 810px) {
         top: -7rem;
         left: 0%;
